backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.45
python-dotenv==1.2.1
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9